Initial restructuring
I have removed a lot of files, to avoid build errors and unnecessary
complication. This gives a room for adding the files back one at a time
and finding best place to place them.

I don't have a clear view on what this will look at the end. Probably it
will change beyond recognition but I intend to keep the API.